[image] refactor(getDefaults) moved to a separate module
