+++ packages/plugin-image/src/consts.js
@@ -1,0 +1,10 @@
+export const getDefault = () => ({
+  config: {
+    alignment: 'center',
+    size: 'content',
+    showTitle: true,
+    showDescription: true,
+  },
+});
+
+export const SMALL_SIZE_WIDTH = 350;
