+++ packages/plugin-image/src/index.js
@@ -1,4 +1,4 @@
 export { createImagePlugin } from './createImagePlugin';
-export { ImageViewer } from './image-viewer';
+export { default as ImageViewer } from './image-viewer';
 export { IMAGE_TYPE_LEGACY, IMAGE_TYPE } from './types';
 export { Modals, ModalsMap } from './modals';
